+++ src/vbox/safe_array.hpp
@@ -18,7 +18,7 @@ struct SafeArray {
 	SafeArray(SafeArray&& other);
 	SafeArray& operator=(SafeArray&& other);
 
-	void copy_in(void* data, size_t size);
+	void copy_in(void* data, ULONG size);
 	ArrayOut copy_out(VARTYPE vartype);
 	ArrayOut copy_out();
